requirements.txt
django
djangorestframework
psycopg2-binary
django-extensions
pyyaml
djangorestframework
django-rest-auth
